pipeline/requirements.txt
azure-core==1.29.0
azure-storage-blob==12.17.0
azure-storage-file-share==12.13.0
click==8.1.6
pandas==1.3.2
python-dotenv==0.19.2
requests==2.31.0
twilio==7.3.1
openpyxl==3.0.9